:bulb: Feat: "Detail 페이지"
fist commit
merge

Resolves: #61
