--- src/main/java/ultimatum/project/global/config/BeanConfig.java
@@ -1,14 +1,0 @@
-package ultimatum.project.global.config;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class BeanConfig {
-
-    @Bean
-    public ModelMapper modelMapper() {
-        return new ModelMapper();
-    }
-}
