--- src/main/java/ultimatum/project/global/config/WebConfig.java
@@ -1,17 +1,0 @@
-package ultimatum.project.global.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebConfig implements WebMvcConfigurer {
-
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/api/**")
-                .allowedOrigins("http://localhost:8083")  //vue localhost 주소
-                .allowedMethods("GET","POST", "PUT", "DELETE");
-    }
-
-}
