+++ src/main/java/ultimatum/project/repository/event/RecommendListEventRepository.java
@@ -3,6 +3,7 @@ package ultimatum.project.repository.event;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import ultimatum.project.domain.entity.event.RecommendListEvent;
 import ultimatum.project.domain.entity.food.RecommendListFood;
